requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
sqlmodel>=0.0.8
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
psycopg2-binary>=2.9.5
sqlalchemy>=1.4.0